requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pyTelegramBotAPI==3.6.7
python-dotenv==0.11.0
requests==2.31.0
six==1.16.0
urllib3==2.0.4
